% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_extract.R
\name{reg_extract}
\alias{reg_extract}
\title{Extract regression coefficients}
\usage{
reg_extract(model)
}
\arguments{
\item{model}{An lm or glm object}
}
\value{
A data.frame with columns: term, estimate, std.error, statistic, p.value
}
\description{
Extracts coefficients, standard errors, t/z values, and p-values from
an lm or glm object and returns a tidy data.frame.
}
\examples{
lm_model <- lm(mpg ~ wt + hp, data = mtcars)
reg_extract(lm_model)

}
