% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_extract.R
\name{reg_plot}
\alias{reg_plot}
\title{Plot regression coefficients}
\usage{
reg_plot(model, show_p = TRUE)
}
\arguments{
\item{model}{An lm or glm object}

\item{show_p}{Logical, whether to highlight significant coefficients (p < 0.05)}
}
\value{
A ggplot object
}
\description{
Creates a forest-style plot of regression coefficients with 95\% CI.
}
\examples{
lm_model <- lm(mpg ~ wt + hp, data = mtcars)
reg_plot(lm_model)

}
